fix(count): avoid shift-by-32 UB when newline lands on bit 31
count_lines_with_pattern cleared processed candidates with
`1u << (newline_bit + 1)`; when ctz placed newline on bit 31 the shift
exponent reached 32 which is undefined behaviour on a 32-bit unsigned.
Use a uint64 shift so the +1 always fits before narrowing back.

Regression test added with a single newline at byte 31 of the first
SIMD chunk and a candidate at byte 0; triggers the UB under UBSan
before the fix, passes after.

Found via libFuzzer harness fuzz_bc_core_search.

+++ src/memory/bc_core_count.c
@@ -234,7 +234,7 @@ __attribute__((target("avx2"))) static bool bc_core_count_lines_with_pattern_avx
                 count++;
             }
             current_line_has_match = false;
-            candidate_mask &= ~((1u << (newline_bit + 1)) - 1);
+            candidate_mask &= (unsigned int)~(((uint64_t)1 << (newline_bit + 1)) - 1);
             newline_mask &= newline_mask - 1;
         }
